+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## (Unreleased)
+## 3.4.1 October 20, 2017
 
 - patch for manual prefixed at-rules. i.e `@-moz-document {...}`.
 - patch for flexbox vendor prefix `flex-grow`, `flex-shrink` and `flex-basis`.